extract OfferItem from OfferList
