+++ src/pages/Browse/TaskDetail/Offers/OfferList.jsx
@@ -3,36 +3,8 @@ import React, { useContext } from 'react';
 import styles from './OfferList.module.scss';
 
 import { TaskContext } from '../TaskContext';
-import CommentAvatar from '../Comment/CommentAvatar';
 import EmptyOffer from './EmptyOffer';
-import useCollapsible from './useCollapsible';
-
-function OfferItem({ offer }) {
-  const Collapsible = useCollapsible();
-
-  const {
-    offeredBy: {
-      avatar, fullName,
-    },
-    message,
-  } = offer;
-
-  return (
-    <div className={styles.offer_wrapper} >
-      <div className={styles.header} >
-        <CommentAvatar avatarUrl={avatar} />
-        <div className={styles.name} >
-          {`${fullName}`}
-        </div>
-      </div>
-      <Collapsible>
-        <div className={styles.content}>
-          {message}
-        </div>
-      </Collapsible>
-    </div>
-  );
-}
+import OfferItem from './OfferItem';
 
 export default function OfferList() {
   const { offers } = useContext(TaskContext);
